+++ grobid-service/src/main/java/org/grobid/service/GrobidRestService.java
@@ -1005,6 +1005,20 @@ public class GrobidRestService implements GrobidPaths {
         return restProcessTraining.resultTraining(token);
     }
 
+    @Path(PATH_ALL_TRAINING)
+    @Produces(MediaType.APPLICATION_JSON)
+    @GET
+    public Response allTraining() {
+        return restProcessTraining.allTraining();
+    }
+
+    @Path(PATH_KILL_TRAINING)
+    @Produces(MediaType.APPLICATION_JSON)
+    @DELETE
+    public Response killTraining(@QueryParam("token") String token) {
+        return restProcessTraining.killTraining(token);
+    }
+
     @Path(PATH_MODEL)
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     @Produces("application/zip")